+++ library/src/main/java/com/bumptech/glide/manager/ConnectivityStrategy.java
@@ -1,0 +1,10 @@
+package com.bumptech.glide.manager;
+
+/** A strategy that to perceive connectivity changes. */
+public interface ConnectivityStrategy {
+  /** perceive connectivity changes from Framework. */
+  void register();
+
+  /** stop perceive connectivity changes from Framework. */
+  void unregister();
+}
