+++ library/src/main/java/com/bumptech/glide/manager/DefaultConnectivityMonitorFactory.java
@@ -2,6 +2,8 @@ package com.bumptech.glide.manager;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
+import android.os.Build.VERSION;
+import android.os.Build.VERSION_CODES;
 import android.util.Log;
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
@@ -29,8 +31,17 @@ public class DefaultConnectivityMonitorFactory implements ConnectivityMonitorFac
               ? "ACCESS_NETWORK_STATE permission granted, registering connectivity monitor"
               : "ACCESS_NETWORK_STATE permission missing, cannot register connectivity monitor");
     }
-    return hasPermission
-        ? new DefaultConnectivityMonitor(context, listener)
-        : new NullConnectivityMonitor();
+
+    ConnectivityStrategy strategy;
+    if (hasPermission) {
+      if (VERSION.SDK_INT >= VERSION_CODES.M) {
+        strategy = new PostMConnectivityStrategyImpl(context, listener);
+      } else {
+        strategy = new PreMConnectivityStrategyImpl(context, listener);
+      }
+    } else {
+      strategy = new NullConnectivityStrategyImpl();
+    }
+    return new DefaultConnectivityMonitor(strategy);
   }
 }
